♻️ Moved ChartProductionData to page folder
