+++ frontend/src/pages/ProductionPage/ChartProductionData.jsx
@@ -18,14 +18,14 @@ import minMax from 'dayjs/plugin/minMax'
 import Papa from 'papaparse'
 import Chart from '@somenergia/somenergia-ui/Chart'
 import SumDisplay from '@somenergia/somenergia-ui/SumDisplay'
-import PageTitle from './PageTitle'
-import ContractSelector from './ContractSelector'
-import Loading from './Loading'
-import ovapi from '../services/ovapi'
-import format from '../services/format'
-import { index2time, timeSlice, sliceIndexes } from '../services/curves'
-import { downloadTextFile } from '../services/download'
-import { InstallationContext } from './InstallationProvider'
+import PageTitle from '../../components/PageTitle'
+import Loading from '../../components/Loading'
+import ContractSelector from '../../components/ContractSelector'
+import { InstallationContext } from '../../components/InstallationProvider'
+import ovapi from '../../services/ovapi'
+import format from '../../services/format'
+import { index2time, timeSlice, sliceIndexes } from '../../services/curves'
+import { downloadTextFile } from '../../services/download'
 
 dayjs.extend(minMax)
