+++ frontend/src/pages/ProductionPage/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import Container from '@mui/material/Container'
-import ChartProductionData from '../../components/ChartProductionData'
+import ChartProductionData from './ChartProductionData'
 
 export default function ProductionPage(params) {
   const { t } = useTranslation()
